requirements.txt
awesome-slugify==1.6.5
cssmin==0.2.0
Flask==0.11.1
Flask-Analytics==0.5.5
Flask-Assets==0.11
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Misaka==0.4.1
Flask-SQLAlchemy==2.1
Flask-User==0.6.8
Flask-WaffleConf==0.3.1
Flask-WTF==0.12
